refactoring
Moving more of the graphics code to the graphics.js   Using canvas.width  iinstead of the $(#wcanvas).width() routines
